+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### General Changes
 * Static builds now based on latest Alpine Linux 3.x instead of always 3.14.
+* Names of discovered files/dirs now get printed before other processing, such as copy or unlink.
 
 ### Contributors
 * Thanks to Ido Szargel for helpful comments and suggestions.