+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 ### New Features & Enhancements
 * New option "--newer PATH" to filter based on mtime of given path.
 * New option "--exec" to execute arbritrary system commands for each discovered file/dir.
+* New option "--quit" to quit after the first match was found.
 
 ### General Changes
 * Static builds now based on latest Alpine Linux 3.x instead of always 3.14.
 * Names of discovered files/dirs now get printed before other processing, such as copy or unlink.
 
 ### Contributors
-* Thanks to Ido Szargel for helpful comments and suggestions.
+* Thanks to Ido Szargel, Adar Zinger, Deborah Gironde for helpful comments and suggestions.
 
 ### Fixes
 * Fixed potential file descriptor leak when reading dir contents fails.